requirements_clean.txt
absl-py==2.3.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
alembic==1.17.2
altgraph==0.17.4
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.11.0
argcomplete==3.1.4
asgiref==3.10.0
asteroid-filterbanks==0.4.0
asttokens==3.0.0
attrs==25.1.0
audioread==3.0.1
av==15.1.0
Babel==2.10.3
bcrypt==3.2.2
beautifulsoup4==4.13.3
blinker==1.9.0
Brotli==1.1.0
certifi==2025.10.5
cffi==2.0.0
chardet==5.2.0
charset-normalizer==3.4.1
chromedriver-py==141.0.7390.65
click==8.1.8
clipboard==0.0.4
colorama==0.4.6
coloredlogs==15.0.1
colorlog==6.10.1
comm==0.2.2
configobj==5.0.8
contourpy==1.3.3
coverage==5.5
crudini==0.9.4
cryptography==41.0.7
ctransformers==0.2.27
ctranslate2==4.6.0
cv==1.0.0
cycler==0.12.1
datasets==3.6.0
debugpy==1.8.13
decorator==5.2.1
deepdiff==8.6.1
defusedxml==0.7.1
Deprecated==1.2.18
deprecation==2.1.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
Django==5.1.6
dnspython==2.6.1
docx==0.2.4
einops==0.8.1
einx==0.3.0
evdev==1.9.2
ewmh==0.1.6
executing==2.2.1
eyed3==0.9.7
fastapi==0.118.2
fasteners==0.18
faster-whisper==1.2.1
feedparser==6.0.10
ffmpy==0.6.2
filelock==3.20.0
filetype==1.2.0
Flask==3.1.2
flatbuffers==25.9.23
fonttools==4.60.1
frozendict==2.4.6
frozenlist==1.8.0
fsspec==2025.3.0
ftfy==6.3.1
googleapis-common-protos==1.72.0
gradio==5.49.1
gradio_client==1.13.3
greenlet==3.2.4
groovy==0.1.2
grpcio==1.76.0
gTTS==2.5.4
h11==0.16.0
hf-xet==1.2.0
hf_transfer==0.1.9
hidapi==0.14.0.post4
html2text==2024.2.26
httpcore==1.0.9
httplib2==0.20.4
httpx==0.28.1
huggingface-hub==0.36.0
humanfriendly==10.0
hydra-core==1.3.2
HyperPyYAML==1.2.2
hypothesis==6.140.3
idna==3.11
ifaddr==0.2.0
image==1.5.33
imageio==2.37.0
imageio-ffmpeg==0.6.0
IMDbPY==2021.4.18
img2pdf==0.6.1
importlib_metadata==8.7.0
infi==0.0.1
infi.systray==0.1.12.1
iniparse==0.5
inquirerpy==0.3.4
ipykernel==6.29.5
ipython==9.6.0
ipython_pygments_lexers==1.1.1
itsdangerous==2.2.0
jax==0.7.1
jaxlib==0.7.1
jedi==0.19.2
Jinja2==3.1.6
joblib==1.5.2
jsonpatch==1.32
jsonpointer==2.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
julius==0.2.7
jupyter_client==8.6.3
jupyter_core==5.7.2
kenlm==0.3.0
keyboard==0.13.5
kiwisolver==1.4.9
launchpadlib==1.11.0
lazr.restfulclient==0.14.6
lazr.uri==1.0.6
lazy_loader==0.4
libevdev==0.5
libretranslatepy==2.1.1
librosa==0.11.0
lightning==2.5.6
lightning-utilities==0.15.2
llama_cpp_python==0.3.16
llvmlite==0.45.1
lxml==5.2.1
Mako==1.3.10
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.1.7
mdurl==0.1.2
meson==1.3.2
mistral_common==1.8.5
ml_dtypes==0.5.3
monotonic==1.6
more-itertools==10.8.0
MouseInfo==0.1.3
mpmath==1.3.0
msgpack==1.1.2
mss==10.1.0
multidict==6.7.0
multiprocess==0.70.16
mutagen==1.46.0
nest-asyncio==1.6.0
netaddr==0.8.0
netifaces==0.11.0
networkx==3.5
nltk==3.9.2
numba==0.62.1
numpy==2.3.5
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.17.0
opencv-contrib-python==4.11.0.86
opencv-python==4.11.0.86
opentelemetry-api==1.34.1
opentelemetry-exporter-otlp==1.34.0
opentelemetry-exporter-otlp-proto-common==1.34.0
opentelemetry-exporter-otlp-proto-grpc==1.34.0
opentelemetry-exporter-otlp-proto-http==1.34.0
opentelemetry-proto==1.34.0
opentelemetry-sdk==1.34.1
opentelemetry-semantic-conventions==0.55b1
opt_einsum==3.4.0
orderly-set==5.5.0
orjson==3.11.4
packaging==25.0
pandas==2.3.3
parso==0.8.5
pexpect==4.9.0
pfzy==0.3.4
pikepdf==10.0.2
pillow==11.3.0
platformdirs==4.5.0
pooch==1.8.2
prompt_toolkit==3.0.52
propcache==0.4.1
protobuf==5.29.5
psutil==7.1.3
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pyarrow==22.0.0
pycountry==24.6.1
pycparser==2.23
pydantic==2.11.10
pydantic-extra-types==2.10.6
pydantic_core==2.33.2
pydub==0.25.1
Pygments==2.19.2
pyparsing==3.2.5
pyperclip==1.11.0
python-dateutil==2.9.0.post0
python-multipart==0.0.20
python-xlib==0.33
python3-xlib==0.15
pytorch-lightning==2.5.6
pytz==2025.2
PyYAML==6.0.3
pyzmq==27.1.0
referencing==0.37.0
regex==2025.11.3
requests==2.32.5
rich==14.2.0
rpds-py==0.29.0
ruamel.yaml==0.18.16
ruamel.yaml.clib==0.2.15
ruff==0.14.6
safehttpx==0.1.7
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
semantic-version==2.10.0
sentencepiece==0.2.1
setuptools==80.9.0
sgmllib3k==1.0.0
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.3
soundfile==0.13.1
soupsieve==2.8
soxr==1.0.0
SQLAlchemy==2.0.44
sqlparse==0.5.3
stack-data==0.6.3
starlette==0.48.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.1
toml==0.10.2
tomlkit==0.13.3
torchmetrics==1.8.2
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
transformers==4.57.3
typer==0.20.0
typer-slim==0.20.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wadllib==2.0.0
wcwidth==0.2.14
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.3
xxhash==3.6.0
yarl==1.22.0
zipp==3.23.0
